pyproject.toml
[tool.poetry]
name = "reddit"
version = "0.1.0"
description = ""
authors = ["Andrei Prădan <andrei@pradan.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
websockets = "^10.2"
python-telegram-bot = "^13.11"
python-dotenv = "^0.19.2"
fastapi = "^0.75.0"
uvicorn = {version = "^0.17.5", extras = ["standard"]}
PyGithub = "^1.55"
pre-commit = "^2.17.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fastapi = 'reddit.web.main:start'
full-setup = 'deploy.tools:full_setup'
set-hook = 'deploy.tools:set_url'
start-services = 'deploy.tools:start_services'
websocket = 'reddit.websocket:start'
